Prophecies system changes

+++ LL/src/Core/Application/UseCases/Leaderboards/Dtos/LeaderboardEntryDto.cs
@@ -9,7 +9,7 @@ public class LeaderboardEntryDto : IMapFrom<LeaderboardEntry>
     public string CharacterName { get; set; } = string.Empty;
     public int Rank { get; set; }
     public int Level { get; set; }
-    public int Experience { get; set; }
+    public long Experience { get; set; }
     public void Mapping(Profile profile)
     {
         profile.CreateMap<LeaderboardEntry, LeaderboardEntryDto>();
